requirements.txt
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
click==7.1.2
colorama==0.4.4
cryptography==3.3.1
decorator==4.4.2
Django==3.1.3
django-crispy-forms==1.10.0
django-filter==2.4.0
django-rest-knox==4.1.0
djangorestframework==3.12.2
dynaconf==3.1.2
flake8==3.8.4
hvac==0.10.5
idna==2.10
isort==5.6.4
knox==0.1.11
lazy-object-proxy==1.4.3
loguru==0.5.3
mccabe==0.6.1
Pillow==8.0.1
psycopg2==2.8.6
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pytz==2020.4
requests==2.25.0
shortuuid==1.0.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.2
validators==0.18.1
win32-setctime==1.0.3
wrapt==1.12.1